Move SDK to subfolder for packaging
